Upgrade Semantic-UI-React to 0.61.0
I was previously on 0.60.9.  Unfortunately, that apparently had a huge
performance issue, due to the way Lodash was being used to omit props
that components didn't handle themselves.  That resulted in increasingly
slow performance to switch tabs as I added more components per tab
panel (and not that many components, either!).  The perf issue was
fixed in 0.60.10, naturally.

References:

https://github.com/Semantic-Org/Semantic-UI-React/pull/860
https://github.com/Semantic-Org/Semantic-UI-React/issues/864